# Streaming Symbol Parser