… Cargo.lock
… breakpad-symbols/src/sym_file/parser.rs