This changes the breakpad-symbols parser to stream its input. This allows us to:

* Avoid ever manifesting the entire (~GB) symbol file in memory
* Parse as the file downloads
* Save to the cache as the file downloads
* Theoretically eliminate a ton of (re)allocations and copies

See breakpad_symbols/src/syms/mod.rs : SymbolFile::parse for discussion of the design