@Gankra 2021-12-14
I haven't actually benchmarked/measured this at all. I'm not great at that stuff so if anyone wants to volunteer...