@Gankra 2021-12-15
shoutouts to the ~worst~ best symbol in firefox 

https://twitter.com/Gankra_/status/1470917063220899848?s=20

![image](https://user-images.githubusercontent.com/1136864/146102593-711c91a3-9c61-4cc8-8778-cbcd1456ef04.png)