@Gankra 2021-12-15
note to self, this may be useful: https://github.com/rust-bakery/nom-bufreader