@Gankra 2021-12-16
I taught socc-pair how to do some rudimentry benchmarking (hot cache, didn't want to download 20+gb of symbol files on wifi, but if you pass `--clean-cache` it will redownload the symbols every time):

```
socc-pair --api-token=... --run-local=/Users/ABeingessner/dev/rust-minidump/minidump-stackwalk/ --crash-id=082a152f-05b4-4de9-943b-c30550211210 --bench=20 --skip-diff
```

Before:

```
benchmark results (ms):
  2549, 2271, 2194, 2226, 2190, 2264, 2194, 2199, 2202, 2211, 2202, 2186, 2196, 2202, 2199, 2224, 2175, 2232, 2192, 2190, 
average runtime: 00m:02s:0224ms
median runtime: 00m:02s:0202ms
min runtime: 00m:02s:0175ms
max runtime: 00m:02s:0549ms
```

After:

```
benchmark results (ms):
  2111, 2041, 2106, 2068, 2188, 2130, 2093, 2099, 2255, 2132, 2112, 2036, 2062, 2066, 2063, 2048, 2040, 2058, 2061, 2059, 
average runtime: 00m:02s:0091ms
median runtime: 00m:02s:0068ms
min runtime: 00m:02s:0036ms
max runtime: 00m:02s:0255ms
```

Summary:

For a hot cache (symbol files are on disk), this shaves about 200ms off of a 2200ms average runtime (9% reduction in runtime). Similar reductions are seen for min, max, and median.

Haven't measured how memory usage is affected (not sure how to do that portably from rust code), but I have to assume it's way lower?

Anyone who wants to run the --clean-cache version and destroy their network, feel free.