@luser 2021-12-16
> Haven't measured how memory usage is affected (not sure how to do that portably from rust code), but I have to assume it's way lower?

I was going to suggest a non-portable Linux-only solution using `wait4(2)` but a cursory search discovered that someone has already made a lovely Rust crate that wraps it *and* provides the functionality in a cross-platform way:
https://docs.rs/wait4/latest/wait4/