@luser 2021-12-16
On Linux it might also be interesting to run them under `perf` and count the number of calls to `malloc` or the total bytes allocated via `malloc`:
https://stackoverflow.com/a/52732823/69326