@luser 2021-12-16
> Anyone who wants to run the --clean-cache version and destroy their network, feel free.

A more repeatable (and less network-destroying) method of testing would be to run your test once (to populate the cache), then copy the contents of the cache to a different location on disk, serve that directory via a local httpd (I use Python's built-in `http.server` for this a lot: `python3 -m http.server <port> --directory <dir>`), and point mdsw at `localhost` for its symbols.