@luser 2021-12-16
As always I got too excited and down in the weeds with follow-up suggestions so I forgot to state the obvious: nice work!