@Gankra 2021-12-16
My dev machine is windows, so anything with `perf` is strictly a someone-else problem (I also have a macbook pro, but she's getting Old so I don't trust her reliability for profiling).

Wait4 looks interesting though, I'll give that a try.

And yeah maybe I should add a feature to socc-pair to implement #347...