@jrmuizel 2021-12-16
You could try using https://github.com/jrmuizel/etw-profiler/tree/main/etw-gecko to get profiles on Windows. It should work reasonably well.