@Gankra 2021-12-16
Nice, Wait4 worked great on windows! (haven't pushed these changes to socc-pair up yet)

# Summary

Runtimes are all still slightly better with streaming, but the ~infinite resources of my work machine means i/o is an even larger % of runtime, so difference is smaller.

**MAX MEMORY USAGE REDUCED BY 35%** 🎉 (380MB vs 250MB)

Memory seems to be very consistent. We can't get complete order of magnitude changes because ultimately we do still need to construct datastructures that are linear in the size of the symbol files.

I also did one-iteration "cold" runs just to see if anything was weird.
 
We see the same difference in memory usage between the cold cases. Cold runtimes being dominated by downloading (even though I can get them all in under 6 seconds!! holy shit...) means the noise is too high to draw conclusions about time in with only one iteration.

**Still, these seems like a pretty clear slamdunk!**




# Numbers

Testing [9520523c-45ff-4d6c-a186-4fec80211216](https://crash-stats.mozilla.org/report/index/9520523c-45ff-4d6c-a186-4fec80211216)

This time on my beefy windows machine:

* Processor: AMD Ryzen 9 3950X 16-Core Processor 3.50 GHz
* RAM: 32.0 GB (DDR4? 3?)
* Storage: Samsung SSD 970 EVO Plus 1TB
* Network: ~360mb/s wifi

## Serial parser --clean-cache (cold, 1 run)

```
miniump-stackwalk runtime: 00m:06s:718ms (6718ms)
miniump-stackwalk max memory (rss): 385MB (404111360 bytes)
```




## Stream parser --clean-cache (cold, 1 run)

```
miniump-stackwalk runtime: 00m:06s:754ms (6754ms)
miniump-stackwalk max memory (rss): 255MB (268169216 bytes)
```






## Serial parser (hot, 20 runs)

```
benchmark results (ms):
  1378, 1379, 1355, 1345, 1344, 1347, 1342, 1339, 1328, 1337, 1317, 1325, 1381, 1337, 1354, 1327, 1335, 1339, 1334, 1333,
average runtime: 00m:01s:343ms (1343ms)
median runtime: 00m:01s:339ms (1339ms)
min runtime: 00m:01s:317ms (1317ms)
max runtime: 00m:01s:381ms (1381ms)
```

```
max memory (rss) results (bytes):
  400371712, 399323136, 399863808, 398790656, 399015936, 400171008, 400056320, 400539648, 399781888, 400064512, 399171584, 399843328, 399183872, 399147008, 399175680, 400134144, 399515648, 399470592, 400302080, 398868480,
average max-memory: 381MB (399639552 bytes)
median max-memory: 381MB (399781888 bytes)
min max-memory: 380MB (398790656 bytes)
max max-memory: 381MB (400539648 bytes)
```





## Stream parser (hot, 20 runs)

```
benchmark results (ms):
  1245, 1251, 1257, 1260, 1249, 1250, 1255, 1247, 1256, 1261, 1256, 1364, 1260, 1251, 1256, 1245, 1265, 1248, 1258, 1245,
average runtime: 00m:01s:258ms (1258ms)
median runtime: 00m:01s:256ms (1256ms)
min runtime: 00m:01s:245ms (1245ms)
max runtime: 00m:01s:364ms (1364ms)
```

```
max memory (rss) results (bytes):
  263340032, 263106560, 262766592, 262905856, 263315456, 263131136, 263946240, 262561792, 263016448, 262860800, 263168000, 263868416, 263610368, 263434240, 263364608, 263110656, 262639616, 263135232, 263684096, 263581696,
average max-memory: 251MB (263227392 bytes)
median max-memory: 250MB (263168000 bytes)
min max-memory: 250MB (262561792 bytes)
max max-memory: 251MB (263946240 bytes)
```