@Gankra 2021-12-16
Just for fun I went back to before #357 to see what things were like (should only affect cold download):

```
miniump-stackwalk runtime: 00m:09s:610ms (9610ms)
miniump-stackwalk max memory (rss): 397MB (417112064 bytes)
```

Runtime is noise, but looks like that PR shaved off like 4% peak memory usage. Maybe more dramatic for bigger symbols. (not sure what the most egregious XUL.syms are)