@Gankra 2021-12-17
I'm happy enough with this, let's land it.